requirements.txt
alembic==1.0.7
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
Mako==1.0.7
MarkupSafe==1.1.0
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
redis==3.1.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
